% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NA_row.R
\name{create_NA_row}
\alias{create_NA_row}
\title{Create NA row}
\usage{
create_NA_row(df)
}
\arguments{
\item{df}{The empty data frame name.}
}
\description{
Create a row in empty data frame with 'NA'.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
