% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_names.R
\name{get_file_names}
\alias{get_file_names}
\title{List file names in folder}
\usage{
get_file_names(filepath, pattern = ".tsv", recursive = TRUE)
}
\arguments{
\item{filepath}{The path to the folder}

\item{pattern}{The pattern to search for}

\item{recursive}{Logical, search in sub-folders or not?}
}
\description{
List file names in folder with pattern and recursive settings.
}
\author{
Håkon Kaspersen, \email{hakon.kaspersen@vetinst.no}
}
