% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_import_call.R
\name{make_import_call}
\alias{make_import_call}
\title{Make an import call object.}
\usage{
make_import_call(name, pkg, ns, new_name)
}
\arguments{
\item{name}{character: The name of the object to import.}

\item{pkg}{character: The name of the package to import from.}

\item{ns}{character: The namespace in which to import.}

\item{new_name}{character: The new name to use for the imported object.}
}
\value{
A call object.
}
\description{
The import call constructed by this function can be evaluated with
\code{eval} to perform the actual import.
}

