% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/symbol_as_character.R
\name{symbol_as_character}
\alias{symbol_as_character}
\title{Convert a possible symbol to character.}
\usage{
symbol_as_character(symbol)
}
\arguments{
\item{symbol}{A symbol or character (of length one)}
}
\value{
character
}
\description{
Convert a possible symbol to character.
}

