% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/symbol_list.R
\name{symbol_list}
\alias{symbol_list}
\title{Process a symbol list provided passed as ellipses}
\usage{
symbol_list(...)
}
\arguments{
\item{...}{dot arguments as passed to}
}
\value{
A named character vector.
}
\description{
This function is used within a function that receives dot-arguments,
and will create a named character vector. It ensures that all
entries are named, and will use the value as name when it is missing.
}

