% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import.R
\docType{package}
\name{import}
\alias{import}
\alias{import-package}
\title{An Import Mechanism for R}
\description{
The package provides an alternative mechanism for importing
objects from packages.
}
\details{
import
}
\author{
Stefan Milton Bache
}

