% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_import_call.R
\name{make_import_call}
\alias{make_import_call}
\title{Make an import call object.}
\usage{
make_import_call(params, exports_only)
}
\arguments{
\item{params}{list of parameters to substitute in the call.}

\item{exports_only}{logical indicating whether only exported objects are
allowed.}
}
\value{
A call object.
}
\description{
The import call constructed by this function can be evaluated with
\code{eval} to perform the actual import.
}
